# feat: add async bulk operations with concurrent execution