## Summary

This PR adds async bulk operations to the Bloomy Python SDK, enabling efficient concurrent processing of multiple API requests with configurable rate limiting. Performance benchmarks show 77%+ improvements over sequential operations.

## Key Changes

### New Features
- **Async bulk create operations** (`create_many`) for:
  - Todos
  - Issues  
  - Goals
  - Meetings
- **Async bulk read operation** (`get_many`) for meetings
- **Configurable rate limiting** with `max_concurrent` parameter
- **Best-effort execution** with detailed error tracking via `BulkCreateResult`

### Bug Fixes
- Fixed todo creation endpoint for meetings: `todo/createmeetingtodo` → `L10/{meeting_id}/todos`
- Fixed meeting attendees field mapping: `Id` → `UserId`
- Fixed meeting issues field mappings to match Issue model

### Documentation
- Added comprehensive bulk operations guide with code examples
- Included real-world performance benchmarks
- Documented error handling patterns and best practices

### Testing
- Added extensive test coverage for all bulk operations
- Tests verify concurrent execution behavior
- Tests ensure proper error handling for partial failures

## Performance Impact

Real-world benchmarks show significant improvements:
- Creating 10 todos: 1.97s → 0.45s (77% faster)
- Creating 10 issues: 2.03s → 0.31s (85% faster)
- Creating 10 goals: 2.08s → 0.35s (83% faster)

## Breaking Changes

None. All changes are additive and maintain backward compatibility.

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>